utils.py to include common code and fixed merged
